GenAI/resume-analyzer/requirements.txt
streamlit
langchain
chromadb
huggingface_hub
sentence-transformers
python-dotenv
requests
llama_index
langchain-community
pymupdf
langchain-huggingface
torch
transformers
langchain-chroma

